Change edit mode navigation to be tabbed (#973)
* Change edit mode navigation to be tabbed

* Typo fix

* Updates to imports and comment fixes
