+++ src/apps/ReleaseControl/components/Form/Inputs/index.ts
@@ -2,3 +2,5 @@ export * from './Description/Description';
 export * from './Title/Title';
 export * from './PlannedDueDate/PlannedDueDate';
 export * from './References/References';
+export * from './Scope/Tags';
+export * from './Scope/HtCables';
